Let the load generators scale linear
The load generator divided the data volume by three and rounded the result.
Thus, e.g., for load 2, 3, and 4 the same number of iterations are
performed and result in the same produced data rate.
Now the data volume is taken and not divided.
Further, the data volume starts at one to produce data to have a
ratio scale and data volume of 0 is the absolute zero.